requirements.txt
azure_storage==0.37.0
beautifulsoup4==4.10.0
Flask==2.0.2
MarkupSafe==2.0.1
newsapi==0.1.1
newsapi_python==0.2.6
python-dotenv==0.19.2
requests==2.26.0
sentry_sdk==1.4.3
websocket_client==1.2.1
yfinance==0.1.64
